move follow button under the avatar
